+++ CHANGES.md
@@ -6,6 +6,7 @@
 - FIX: Raise missing file not found exception - by @emmanuel-ferdman
 - FIX: Prune also `driver` keyword in `utils.read`
 - DOC: Fix jupyter cache for documentation notebook [#211](https://github.com/sertit/eoreader/issues/211)
+- DOC: Convert gfm admonitions to rst
 
 ## 0.22.3 (2025-05-05)